Move App.java to under assignmnet folder Modify some braces for readabilty
