+++ app/src/main/java/com/codesoom/assignment/App.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.models;
+package com.codesoom.assignment;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
