+++ app/src/main/java/com/codesoom/assignment/controllers/TaskController.java
@@ -1,6 +1,5 @@
 package com.codesoom.assignment.controllers;
 
-
 import com.codesoom.assignment.models.Task;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,18 +14,27 @@ public class TaskController {
     private Long newId = 0L;
 
     @GetMapping
-    public List<Task> list()
-    {
+    public List<Task> list() {
         return tasks;
     }
 
     @PostMapping
-    public Task create(@RequestBody Task task)
-    {
+    public Task create(@RequestBody Task task) {
         task.setId(generateId());
         tasks.add(task);
         return task;
     }
+
+//    public Task update(@RequestBody Task task)
+//    {
+//        task.setTitle(task.getTitle());
+//        return task;
+//    }
+//    public Task delete(@RequestBody Task task)
+//    {
+//        tasks.remove(task);
+//        return task;
+//    }
 
     private Long generateId() {
         newId++;
