+++ app/src/test/java/com/codesoom/assignment/AppTest.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment;
 
-import com.codesoom.assignment.models.App;
+import com.codesoom.assignment.App;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
